refactor(cockpit): extract review packet manifest
